requirements.txt
blis==0.7.4
boto3==1.17.76
botocore==1.20.76
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
dill==0.3.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
fastapi==0.65.1
idna==2.10
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.20.3
packaging==20.9
pathy==0.5.2
preshed==3.0.5
pydantic==1.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
requests==2.25.1
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
starlette==0.14.2
thinc==8.0.3
threadpoolctl==2.1.0
tqdm==4.60.0
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wasabi==0.8.2
zipp==3.4.1
